Refactor consume and test error handling
