+++ batch-processing/main_test.go
@@ -2,6 +2,7 @@ package main
 
 import (
 	"encoding/csv"
+	"errors"
 	"reflect"
 	"strings"
 	"testing"
@@ -68,5 +69,52 @@ func TestConsume(t *testing.T) {
 		if calls != len(testCase.ExpectedRows) {
 			t.Fatalf("%v: incorrect num calls to row processing function, expected %v, got %v", testCase.Name, len(testCase.ExpectedRows), calls)
 		}
+	}
+}
+
+func TestHeaderError(t *testing.T) {
+	r := csv.NewReader(strings.NewReader("a\nb"))
+
+	expected := "inner header error"
+	err := consume(r, func(s []string) error {
+		return errors.New(expected)
+	}, func(s []string) error {
+		t.Fatal("row processing function called unexpectedly")
+		return nil
+	})
+
+	if err == nil {
+		t.Fatal("consume did not return an error")
+	}
+
+	inner := errors.Unwrap(err)
+	if inner == nil {
+		t.Fatal("incorrect error: could not unwrap")
+	}
+	if inner.Error() != expected {
+		t.Fatalf("incorrect error: expected %v, got %v", expected, inner.Error())
+	}
+}
+
+func TestRowError(t *testing.T) {
+	r := csv.NewReader(strings.NewReader("a\nb"))
+
+	expected := "inner row error"
+	err := consume(r, func(s []string) error {
+		return nil
+	}, func(s []string) error {
+		return errors.New(expected)
+	})
+
+	if err == nil {
+		t.Fatal("consume did not return an error")
+	}
+
+	inner := errors.Unwrap(err)
+	if inner == nil {
+		t.Fatal("incorrect error: could not unwrap")
+	}
+	if inner.Error() != expected {
+		t.Fatalf("incorrect error: expected %v, got %v", expected, inner.Error())
 	}
 }
